+++ src/main/java/org/sopt/post/application/PostService.java
@@ -4,11 +4,11 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.sopt.post.application.request.CreatePostServiceRequest;
 import org.sopt.post.domain.Tag;
-import org.sopt.post.presentation.request.UpdatePostRequest;
-import org.sopt.post.presentation.response.GetAllPostsResponse;
-import org.sopt.post.presentation.response.GetDetailedPostResponse;
-import org.sopt.post.presentation.response.GetSimplePostResponse;
-import org.sopt.post.presentation.response.SearchResultResponse;
+import org.sopt.post.presentation.dto.request.UpdatePostRequest;
+import org.sopt.post.presentation.dto.response.GetAllPostsResponse;
+import org.sopt.post.presentation.dto.response.GetDetailedPostResponse;
+import org.sopt.post.presentation.dto.response.GetSimplePostResponse;
+import org.sopt.post.presentation.dto.response.SearchResultResponse;
 import org.sopt.user.domain.User;
 import org.sopt.user.presentation.exception.UserErrorCode;
 import org.sopt.user.infrastructure.repository.UserRepository;
