+++ src/main/java/org/sopt/post/presentation/dto/response/GetSimplePostResponse.java
@@ -1,11 +1,11 @@
-package org.sopt.post.presentation.response;
+package org.sopt.post.presentation.dto.response;
 
 import org.sopt.post.domain.Post;
 import org.sopt.user.presentation.dto.response.GetUserResponse;
 
-public record SearchPostResponse(Long id, GetUserResponse author, String title) {
-    public static SearchPostResponse from(Post post) {
-        return new SearchPostResponse(
+public record GetSimplePostResponse(Long id, GetUserResponse author, String title) {
+    public static GetSimplePostResponse from(Post post) {
+        return new GetSimplePostResponse(
                 post.getId(),
                 GetUserResponse.from(post.getUser()),
                 post.getTitle()
